\docType{data}
\name{asthma}
\alias{asthma}
\title{SNP data on asthma case-control study}
\format{The \code{asthma} data frame has 1578 rows (individuals) and 57 columns (variables) of data from a genetic study on asthma.}
\usage{
data("asthma")
}
\value{
An \code{data.frame} object.
}
\description{
\code{data.frame} with 51 SNPs and 6 epidemiological variables: country, gender, age, bmi, smoke ans case/control status.
}
\examples{
data(asthma)
dim(asthma)
str(asthma)
}
\keyword{datasets}
