% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneSymbol.R
\name{getGeneSymbol}
\alias{getGeneSymbol}
\title{Get gene symbol from a list of SNPs}
\usage{
getGeneSymbol(
  x,
  snpCol = 1,
  chrCol = 2,
  posCol = 3,
  db = "TxDb.Hsapiens.UCSC.hg19.knownGene"
)
}
\arguments{
\item{x}{data.frame containing: SNP name, chromosome and genomic position.}

\item{snpCol}{column of x having the SNP name. Default is 1.}

\item{chrCol}{column of x having the SNP chromosome. Default is 2.}

\item{posCol}{column of x having the SNP position. Default is 3.}

\item{db}{reference genome. Default is 'TxDb.Hsapiens.UCSC.hg19.knownGene'}
}
\value{
a data.frame having initial information and gene symbol
}
\description{
The getGeneSymbol function searches for the genes associated with the SNPs
for a given chromosome at a given position. To perform the annotation this
function needs to have the BioConductor libraries: S4Vectors, GenomicRanges,
IRanges, VariantAnnotation and  org.Hs.eg.db libraries.
As well as the library with the reference genome database, by default needs
TxDb.Hsapiens.ICSC.hg19.KnownGene library
}
