% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocusInfo.R
\name{getLocusInfo}
\alias{getLocusInfo}
\title{Get info from snpStats objects to be passed to Locus Zoom}
\usage{
getLocusInfo(x, annotation, file, candidate, size = 1e+05)
}
\arguments{
\item{x}{an object of class 'GlmTests'}

\item{annotation}{a data.frame having annotation (obtained from PLINK files)}

\item{file}{saved file}

\item{candidate}{candidate SNPs (top one)}

\item{size}{window size}
}
\value{
file having the required information of Locus Zoom
}
\description{
Get info from snpStats objects to be passed to Locus Zoom
}
