\name{HapMap}
\alias{HapMap}
\docType{data}

\title{ SNPs from HapMap project}

\description{
Information about 9307 SNPs from the HapMap project
belonging to 22 chromosomes. Information about two different population is available:
European population (CEU) and Yoruba (YRI). The genomic
information (names of SNPs, chromosomes and genetic position) is
also available in a data frame called 'HapMap.SNPs.pos'.
 
}

\usage{data(HapMap)}

\format{
  A data frame with 120 observations on the 9808 variables (SNPs) and one variable
  called 'group' indicating the population.
  
}

\source{
  HapMap project (http://www.hapmap.org)
}

\examples{
data(HapMap)
}

\keyword{datasets}
