% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related.R
\name{related}
\alias{related}
\title{Get related samples}
\usage{
related(x)
}
\arguments{
\item{x}{An object obtained from SNPrelate package.}
}
\value{
A matrix with related individuals.
}
\description{
Get related samples
}
\examples{
  library(SNPassoc)
  data(SNPs)

}
