% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPassoc-package.R
\docType{package}
\name{SNPassoc-package}
\alias{SNPassoc}
\alias{SNPassoc-package}
\title{SNPassoc: SNPs-Based Whole Genome Association Studies}
\description{
Functions to perform most of the common analysis in genome association studies are implemented. These analyses include descriptive statistics and exploratory analysis of missing values, calculation of Hardy-Weinberg equilibrium, analysis of association based on generalized linear models (either for quantitative or binary traits), and analysis of multiple SNPs (haplotype and epistasis analysis). Permutation test and related tests (sum statistic and truncated product) are also implemented. Max-statistic and genetic risk-allele score exact distributions are also possible to be estimated. The methods are described in Gonzalez JR et al., 2007 \doi{ 10.1093/bioinformatics/btm025}. This version includes internal copies of functions from the archived 'haplo.stats' package to maintain functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/isglobal-brge/SNPassoc}
}

}
\author{
\strong{Maintainer}: Dolors Pelegri \email{dolors.pelegri@isglobal.org} (\href{https://orcid.org/0000-0002-5993-3003}{ORCID})

Authors:
\itemize{
  \item Victor Moreno
  \item Juan R Gonzalez \email{juanr.gonzalez@isglobal.org} (\href{https://orcid.org/0000-0003-3267-2146}{ORCID})
}

Other contributors:
\itemize{
  \item Vitaly Korchagin \email{vitaly_korchagin@rambler.ru} (\href{https://orcid.org/0000-0003-2264-6294}{ORCID}) [contributor]
}

}
\keyword{internal}
