\name{SNPs.info.pos}
\alias{SNPs.info.pos}

\title{SNPs in a case-control study}
\usage{data(SNPs.info.pos)}
\description{
    SNPs data.frame contains selected SNPs and other clinical covariates for cases and controls 
    in a case-control study
    
    SNPs.info.pos data.frame contains the names of the SNPs included in the data set 'SNPs' including
    their chromosome and their genomic position
}
\format{
    'SNPs.info.pos' data.frame contains the following columns:
        A data frame with 35 observations on the following 3 variables.
    \describe{
        \item{\code{snp}}{name of SNP}
        \item{\code{chr}}{name of chromosome}
        \item{\code{pos}}{genomic position}
    }
}

\source{
    Data obtained from our department. The reference and details will be supplied after being published.
}

\keyword{datasets}
