\name{SNPs.info.pos}
\alias{SNPs.info.pos}
\docType{data}
\title{ Genomic position for selected SNPs}
\description{
  This data.frame contains information about the SNPs regarding their
  genomic position and the gene where they are located for those SNPs 
  in the data.frame \code{\link{SNPs}}. 
}
\usage{data(SNPs.info.pos)}
\format{
  A data frame with 35 observations on the following 3 variables.
  \describe{
    \item{\code{snp}}{name of SNP}
    \item{\code{chr}}{name of chromosome}
    \item{\code{pos}}{genomic position}
  }
}

\source{
  Data obtained from our department. The reference and details will be supplied after being published.
}

\seealso{ \code{\link{SNPs}}}

\keyword{datasets}
