\name{scanWGassociation}
\alias{scanWGassociation}
\title{Whole genome association analysis}
\description{
  This function is obsolete due to some problems with gfotran compiler. 
  Use 'WGassociation' function instead or send an e-mail to the maintainer 
  for receiving a version including this function
}
\keyword{utilities}
